third_party/blink/web_tests/wpt_internal/html/canvas/drawElement/allow-readback/svg-images-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElement with allowReadback does not draw cross-origin SVG images</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <svg id=child width=100 height=200>
      <rect fill="green" x=0 y=0 width=100 height=200 />
      <image id="sameOrigin" href="https://{{location[host]}}/css/css-masking/mask-image/support/red-100x100.png" x=10 y=10 height=100 width=100 />
      <image id="crossOrigin" href="https://{{hosts[alt][www]}}:{{ports[h2][0]}}/css/css-masking/mask-image/support/red-100x100.png" x=0 y=100 height=100 width=100 />
    </svg>
  </canvas>

  <script>
    let sPendingImagesToLoad = 2;

    function pendingImageLoaded() {
      if (!--sPendingImagesToLoad) {
        test(function(t) {
          var context = canvas.getContext("2d");
          context.drawElement(child, 0, 0, { allowReadback: true });

          pixel = context.getImageData(50, 50, 1, 1).data;
          assert_equals(pixel[0], 255, "Same origin should draw");
          assert_equals(pixel[1], 0, "Same origin should draw");
          assert_equals(pixel[2], 0, "Same origin should draw");

          pixel = context.getImageData(50, 150, 1, 1).data;
          assert_equals(pixel[0], 0, "Cross origin should not draw");
          assert_equals(pixel[1], 128, "Cross origin should not draw");
          assert_equals(pixel[2], 0, "Cross origin should not draw");

          t.done();
        });
      }
    }

    sameOrigin.onload = pendingImageLoaded;
    crossOrigin.onload = pendingImageLoaded;
  </script>
</body>
</html>
